school/urls.py
from django.urls import path
from rest_framework.routers import DefaultRouter
from .views import StudentViewSet, CourseViewSet

router = DefaultRouter()
router.register('students', StudentViewSet, basename='student')
router.register('courses', CourseViewSet, basename='course')
urlpatterns = router.urls
